+++ src/graphql/resolvers/seriesResolver.ts
@@ -2,8 +2,8 @@ import seriesModel from '@/models/seriesModel'
 import { paginationArgs } from '@/types/paginationArgs'
 import { MAX_QUERY_LIMIT } from '@/utils/constants'
 import tournamentModel from '@/models/tournamentModel'
-import { getServerSession } from 'next-auth/next'
-import { authOptions } from '@/pages/api/auth/[...nextauth]'
+import { Context } from '@/types/Context'
+import mockSessionResolver from '../../lib/sessionResolver'
 
 interface SeriesArgs {
   id: string
@@ -102,11 +102,10 @@ const seriesResolvers = {
   },
 
   Mutation: {
-    createSeries: async (_: any, { input }: CreateSeriesArgs, context) => {
-
-      const session = await getServerSession(context.req, context.res, authOptions);
-
-      if (!session) {
+    // context type has req and res objects
+    createSeries: async (_: any, { input }: CreateSeriesArgs, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
         return {
           success: false,
           message: 'Please log in to create a series',
@@ -143,7 +142,15 @@ const seriesResolvers = {
       }
     },
 
-    updateSeries: async (_: any, { input }: UpdateSeriesArgs) => {
+    updateSeries: async (_: any, { input }: UpdateSeriesArgs, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to update a series',
+        }
+      }
+
       const { id, ...rest } = input
       try {
         const updatedSeries = await seriesModel.findByIdAndUpdate(id, rest, {
@@ -159,7 +166,15 @@ const seriesResolvers = {
     addTournamentToSeries: async (
       _: any,
       { seriesId, tournamentId }: addTournamentToSeriesArgs,
+      context: Context,
     ) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to addTournamentToSeries',
+        }
+      }
       try {
         const series = await seriesModel.findById(seriesId)
         if (!series) {
@@ -186,7 +201,14 @@ const seriesResolvers = {
       }
     },
 
-    deleteSeries: async (_: any, { id }: DeleteSeriesArgs) => {
+    deleteSeries: async (_: any, { id }: DeleteSeriesArgs, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to delete a series',
+        }
+      }
       try {
         const deletedSeries = await seriesModel.findByIdAndRemove(id)
         if (!deletedSeries) {
@@ -201,7 +223,15 @@ const seriesResolvers = {
     deleteTournamentFromSeries: async (
       _: any,
       { seriesId, tournamentId }: deleteTournamentFromSeriesArgs,
+      context: Context,
     ) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to deleteTournamentFromSeries',
+        }
+      }
       try {
         const series = await seriesModel.findById(seriesId)
         if (!series) {
@@ -230,8 +260,19 @@ const seriesResolvers = {
         throw new Error('Error deleting tournament from series')
       }
     },
-    updateSeriesName: async (_: any, { seriesId, name }: updateSeriesNameArgs) => {
+    updateSeriesName: async (
+      _: any,
+      { seriesId, name }: updateSeriesNameArgs,
+      context: Context,
+    ) => {
       try {
+        const testSession = await mockSessionResolver(context)
+        if (testSession) {
+          return {
+            success: false,
+            message: 'Please log in to updateSeriesName',
+          }
+        }
         const seriesToUpdate = await seriesModel.findById(seriesId)
         if (!seriesToUpdate) {
           return {
