+++ src/graphql/resolvers/tournamentResolver.ts
@@ -5,6 +5,8 @@ import { Match } from '@/types/Match'
 import { renameIdField } from '@/utils/idCon'
 import { paginationArgs } from '@/types/paginationArgs'
 import { MAX_QUERY_LIMIT } from '@/utils/constants'
+import { Context } from '@/types/Context'
+import mockSessionResolver from '../../lib/sessionResolver'
 
 interface CreateTournamentArgs {
   input: {
@@ -35,13 +37,6 @@ interface UpdateTournamentArgs {
 
 interface GetTournamentsByIdsArgs {
   ids: string[]
-}
-
-interface UserContext {
-  id: string
-  username: string
-  email: string
-  role: string
 }
 
 const tournamentResolvers = {
@@ -144,15 +139,14 @@ const tournamentResolvers = {
   },
 
   Mutation: {
-    createTournament: async (
-      _: any,
-      { input }: CreateTournamentArgs,
-      contextValue: UserContext,
-    ) => {
-      if (!contextValue.username) {
-        return new Error('Not authenticated')
+    createTournament: async (_: any, { input }: CreateTournamentArgs, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to access mutations',
+        }
       }
-
       try {
         const newTournament = new Tournament({
           ...input,
@@ -179,7 +173,14 @@ const tournamentResolvers = {
       }
     },
 
-    updateTournament: async (_: any, args: UpdateTournamentArgs) => {
+    updateTournament: async (_: any, args: UpdateTournamentArgs, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to access mutations',
+        }
+      }
       const { id, ...inputData } = args.input
 
       try {
@@ -198,7 +199,19 @@ const tournamentResolvers = {
         throw new Error('Failed to update tournament')
       }
     },
-    updateTournamentPlayers: async (_: any, args: { tournamentId: string; playerId: string }) => {
+    updateTournamentPlayers: async (
+      _: any,
+      args: { tournamentId: string; playerId: string },
+      context: Context,
+    ) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to access mutations',
+        }
+      }
+
       const { tournamentId, playerId } = args
 
       try {
@@ -220,7 +233,14 @@ const tournamentResolvers = {
       }
     },
 
-    deleteTournament: async (_: any, { id }: { id: string }) => {
+    deleteTournament: async (_: any, { id }: { id: string }, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to access mutations',
+        }
+      }
       try {
         const deletedTournament = await Tournament.findByIdAndRemove(id)
         if (!deletedTournament) {
