+++ src/graphql/resolvers/userResolver.ts
@@ -2,6 +2,8 @@ import userModel from '@/models/userModel'
 import bcrypt from 'bcrypt'
 import { paginationArgs } from '@/types/paginationArgs'
 import { MAX_QUERY_LIMIT } from '@/utils/constants'
+import mockSessionResolver from '../../lib/sessionResolver'
+import { Context } from '@/types/Context'
 
 interface UserArgs {
   id: string
@@ -73,7 +75,14 @@ const userResolvers = {
   },
 
   Mutation: {
-    createUser: async (_: any, { input }: CreateUserArgs) => {
+    createUser: async (_: any, { input }: CreateUserArgs, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to access mutations',
+        }
+      }
       try {
         // Hash the password before saving
         const hashedPassword = await bcrypt.hash(input.password, 10)
@@ -91,7 +100,14 @@ const userResolvers = {
       }
     },
 
-    updateUser: async (_: any, { input }: UpdateUserArgs) => {
+    updateUser: async (_: any, { input }: UpdateUserArgs, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to access mutations',
+        }
+      }
       const { id, ...rest } = input
       try {
         if (rest.password) {
@@ -109,7 +125,14 @@ const userResolvers = {
       }
     },
 
-    deleteUser: async (_: any, { id }: DeleteUserArgs) => {
+    deleteUser: async (_: any, { id }: DeleteUserArgs, context: Context) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to access mutations',
+        }
+      }
       try {
         const deletedUser = await userModel.findByIdAndRemove(id)
         if (!deletedUser) {
@@ -121,7 +144,18 @@ const userResolvers = {
         throw new Error('Failed to delete user')
       }
     },
-    sendInvitation: async (_: any, { tournamentId, email }: SendInvitationArgs) => {
+    sendInvitation: async (
+      _: any,
+      { tournamentId, email }: SendInvitationArgs,
+      context: Context,
+    ) => {
+      const testSession = await mockSessionResolver(context)
+      if (testSession) {
+        return {
+          success: false,
+          message: 'Please log in to access mutations',
+        }
+      }
       try {
         const user = await userModel.findOne({ email })
         if (user) {
